bot/basehandlers: be more robust for raw protocol actions
Mainly the change is to check the return value of map lookups. Naiively
a map given an unknown key will return a (nil, false) pair and this was
not expected when it was first written.

This patch fixes all that bad logic and replaces it with proper panics
(so that the recovery mechanism can take over) and better logs known
failure states by dying.